data_collector/requirements.txt
flask
flask[async]==2.2.5
mysql-connector-python
grpcio
grpcio-tools
requests
schedule
confluent-kafka
